refactor: remove duplicate tests from src/server/ - all tests now in tests/python/
🗑️ Deleted:
- src/server/app/tests/ (13 test files)
- src/server/tests/ (full directory with 22 test files)

✅ All tests migrated to:
- tests/python/unit/app/ (13 tests from app/tests)
- tests/python/unit/ (other unit tests)
- tests/python/integration/ (integration tests)

📊 Result:
- No test files in src/ directories
- Single source of truth: tests/python/ (monorepo centralized)
- .pytest_cache removed from index
- src/server now contains ONLY source code
