--- src/server/app/tests/fixtures/__init__.py
@@ -1,3 +1,0 @@
-"""
-Tests: Test fixtures and data factories.
-"""
