--- src/server/app/tests/integration/__init__.py
@@ -1,3 +1,0 @@
-"""
-Tests: Integration tests for API endpoints.
-"""
