--- src/server/app/tests/unit/__init__.py
@@ -1,3 +1,0 @@
-"""
-Tests: Unit tests for core business logic.
-"""
