--- src/server/tests/__init__.py
@@ -1,1 +1,0 @@
-# Empty init file for tests package
