--- src/server/tests/unit/__init__.py
@@ -1,1 +1,0 @@
-# Empty init file for unit tests package
