--- src/server/tests/unit/app/__init__.py
@@ -1,1 +1,0 @@
-# Tests for app module
