--- src/server/tests/unit/app/test_app_coverage.py
@@ -1,195 +1,0 @@
-"""
-Extended tests for app module to increase coverage.
-
-Tests for additional code paths and edge cases.
-"""
-
-from unittest.mock import patch
-
-import pytest
-
-
-class TestAppInitFunctions:
-    """Test app initialization functions."""
-
-    @pytest.mark.asyncio
-    @patch("app.core.database.init_chromadb")
-    @patch("app.core.database.init_sqlite")
-    async def test_lifespan_startup(self, mock_sqlite, mock_chromadb):
-        """✅ Should call initialization functions on startup."""
-        mock_chromadb.return_value = None
-        mock_sqlite.return_value = None
-
-        # Import after patching
-        from app.main import app
-
-        # App should be initialized
-        assert app is not None
-
-    def test_app_routes_configured(self):
-        """✅ Should have proper routes configured."""
-        from app.main import app
-
-        # Should have more than just middleware
-        assert len(app.routes) > 0
-
-    def test_app_exception_handling(self):
-        """✅ App should have exception handlers."""
-        from app.main import app
-
-        # Check exception handlers exist
-        assert hasattr(app, "exception_handlers")
-
-
-class TestAPIEndpointsCoverage:
-    """Test API endpoints for additional coverage."""
-
-    def test_health_router_included(self):
-        """✅ Health router should be included in v1."""
-        from app.api.v1.health import router as health_router
-
-        assert health_router is not None
-
-    def test_knowledge_router_included(self):
-        """✅ Knowledge router should be included in v1."""
-        from app.api.v1.knowledge import router as knowledge_router
-
-        assert knowledge_router is not None
-
-    def test_chat_router_included(self):
-        """✅ Chat router should be included in v1."""
-        from app.api.v1.chat import router as chat_router
-
-        assert chat_router is not None
-
-
-class TestAPIv1RouterIntegration:
-    """Test API v1 router integration."""
-
-    def test_api_v1_router_has_prefixes(self):
-        """✅ Should use correct API prefixes."""
-        from app.api.v1 import router
-
-        # Router should be properly configured
-        assert router is not None
-        assert hasattr(router, "routes")
-
-    def test_multiple_routers_included(self):
-        """✅ Should include multiple sub-routers."""
-        from app.api.v1 import router
-
-        # Should have includes for health, knowledge, chat
-        assert len(router.routes) > 0
-
-
-class TestSecurityFunctions:
-    """Test security module functions."""
-
-    def test_token_validator_validation_logic(self):
-        """✅ TokenValidator should validate API keys."""
-        from app.core.security import TokenValidator
-
-        # Empty key should be invalid
-        assert TokenValidator.validate_api_key("") is False
-
-        # None should be invalid
-        assert TokenValidator.validate_api_key(None) is False
-
-    def test_token_validator_with_valid_format(self):
-        """✅ TokenValidator should accept valid format keys."""
-        from app.core.security import TokenValidator
-
-        # Key with valid format should be tested
-        result = TokenValidator.validate_api_key("valid-api-key-123-456")
-        assert isinstance(result, bool)
-
-
-class TestDatabaseFunctions:
-    """Test database functions."""
-
-    @pytest.mark.asyncio
-    async def test_chromadb_init_callable(self):
-        """✅ init_chromadb should be callable."""
-        from app.core.database import init_chromadb
-
-        # Should be callable function
-        assert callable(init_chromadb)
-
-    @pytest.mark.asyncio
-    async def test_sqlite_init_callable(self):
-        """✅ init_sqlite should be callable."""
-        from app.core.database import init_sqlite
-
-        # Should be callable function
-        assert callable(init_sqlite)
-
-
-class TestConfigurationUsage:
-    """Test configuration is properly used."""
-
-    def test_settings_in_app(self):
-        """✅ Settings should be used in app configuration."""
-        from app.core.config import settings
-        from app.main import app
-
-        # Both should be available
-        assert settings is not None
-        assert app is not None
-
-    def test_config_values_accessible(self):
-        """✅ Config values should be accessible in app."""
-        from app.core.config import settings
-
-        # All config values should be accessible
-        assert settings.APP_NAME is not None
-        assert settings.API_V1_STR is not None
-        assert settings.LOG_LEVEL is not None
-
-
-class TestLoggingSetup:
-    """Test logging is properly configured."""
-
-    def test_logger_configured(self):
-        """✅ Logger should be configured in main."""
-        import logging
-
-        logger = logging.getLogger("app.main")
-        assert logger is not None
-        assert logger.name == "app.main"
-
-    def test_log_level_from_settings(self):
-        """✅ Log level should come from settings."""
-
-        from app.core.config import settings
-
-        # Get current level
-        level = settings.LOG_LEVEL
-        assert level in ["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]
-
-
-class TestEndpointImports:
-    """Test all endpoints can be imported."""
-
-    def test_import_health_endpoint(self):
-        """✅ Should import health endpoint."""
-        from app.api.v1.health import router
-
-        assert router is not None
-
-    def test_import_chat_endpoint(self):
-        """✅ Should import chat endpoint."""
-        from app.api.v1.chat import router
-
-        assert router is not None
-
-    def test_import_knowledge_endpoint(self):
-        """✅ Should import knowledge endpoint."""
-        from app.api.v1.knowledge import router
-
-        assert router is not None
-
-    def test_import_dependencies(self):
-        """✅ Should import dependencies."""
-        from app.api.dependencies import verify_api_key
-
-        assert verify_api_key is not None
