--- src/server/tests/unit/app/test_endpoints_coverage_80.py
@@ -1,133 +1,0 @@
-"""
-Extended tests for app/api/v1 endpoints targeting 80%+ coverage.
-
-Focus on:
-- health.py line 78
-- chat.py line 17
-- knowledge.py line 17
-"""
-
-from fastapi.testclient import TestClient
-
-from app.main import app
-
-
-class TestHealthEndpoint:
-    """Test health check endpoint."""
-
-    def test_health_endpoint_accessible(self):
-        """✅ Health endpoint should be accessible."""
-        client = TestClient(app)
-        response = client.get("/api/v1/health")
-
-        # Should respond (might be 404 if not implemented, but no 500)
-        assert response.status_code < 500
-
-    def test_health_endpoint_returns_json_if_exists(self):
-        """✅ If health endpoint exists, it should return JSON."""
-        client = TestClient(app)
-        response = client.get("/api/v1/health")
-
-        if response.status_code == 200:
-            # If implemented, should be JSON
-            data = response.json()
-            assert isinstance(data, dict)
-
-    def test_health_requires_no_authentication(self):
-        """✅ Health endpoint should not require auth."""
-        client = TestClient(app)
-
-        # Call without API key
-        response = client.get("/api/v1/health")
-
-        # Should not be 401/403
-        assert response.status_code not in [401, 403] or response.status_code == 404
-
-
-class TestChatEndpoint:
-    """Test chat endpoint."""
-
-    def test_chat_endpoint_accessible(self):
-        """✅ Chat endpoint should be accessible or 404."""
-        client = TestClient(app)
-
-        response = client.get("/api/v1/chat")
-
-        # Should respond (not 500)
-        assert response.status_code < 500
-
-    def test_chat_can_handle_requests(self):
-        """✅ Chat endpoint should handle requests gracefully."""
-        client = TestClient(app)
-
-        response = client.get("/api/v1/chat")
-
-        # Anything but 500 is ok
-        assert response.status_code != 500
-
-
-class TestKnowledgeEndpoint:
-    """Test knowledge endpoint."""
-
-    def test_knowledge_endpoint_accessible(self):
-        """✅ Knowledge endpoint should be accessible."""
-        client = TestClient(app)
-
-        response = client.get("/api/v1/knowledge")
-
-        # Should respond (not crash)
-        assert response.status_code < 500
-
-    def test_knowledge_can_handle_requests(self):
-        """✅ Knowledge endpoint should handle requests gracefully."""
-        client = TestClient(app)
-
-        response = client.get("/api/v1/knowledge")
-
-        # Anything but 500 is ok
-        assert response.status_code != 500
-
-
-class TestEndpointIntegration:
-    """Integration tests for all endpoints."""
-
-    def test_api_v1_base_accessible(self):
-        """✅ /api/v1 base path should be accessible."""
-        client = TestClient(app)
-
-        response = client.get("/api/v1")
-
-        # Should not crash
-        assert response.status_code < 500
-
-    def test_api_routes_respond_with_json_or_errors(self):
-        """✅ API routes should respond properly."""
-        client = TestClient(app)
-
-        response = client.get("/api/v1/health")
-
-        # Should be valid HTTP response
-        assert response.status_code >= 200
-
-
-class TestEndpointErrorHandling:
-    """Test endpoint error handling."""
-
-    def test_invalid_endpoint_returns_404(self):
-        """✅ Invalid endpoints should return 404."""
-        client = TestClient(app)
-
-        response = client.get("/api/v1/invalid-endpoint-xyz")
-
-        # Should be 404
-        assert response.status_code == 404
-
-    def test_wrong_method_returns_appropriate_error(self):
-        """✅ Wrong HTTP method should return appropriate error."""
-        client = TestClient(app)
-
-        # Try POST to a GET endpoint if health is GET
-        response = client.post("/api/v1/health")
-
-        # Should be 405 or 404 or similar error
-        assert response.status_code >= 400
