--- src/server/tests/unit/app/test_final_coverage_80.py
@@ -1,214 +1,0 @@
-"""
-Targeted tests for final 80%+ coverage push.
-
-Focus on lines that are simple to cover:
-- main.py lifespan handler yield point
-- Database initialization return values
-- Simple module imports and function existence
-"""
-
-import inspect
-
-from fastapi.testclient import TestClient
-
-from app.main import app, lifespan, shutdown_event, startup_event
-
-
-class TestLifespanContext:
-    """Test lifespan context manager."""
-
-    def test_lifespan_is_async_generator(self):
-        """✅ Lifespan should be async generator function."""
-        # Verify it's a function decorated for lifespan
-        assert callable(lifespan)
-
-    def test_startup_event_is_async_callable(self):
-        """✅ startup_event should be async callable."""
-        assert inspect.iscoroutinefunction(startup_event)
-
-    def test_shutdown_event_is_async_callable(self):
-        """✅ shutdown_event should be async callable."""
-        assert inspect.iscoroutinefunction(shutdown_event)
-
-
-class TestApplicationInitialization:
-    """Test FastAPI app is properly initialized."""
-
-    def test_app_is_fastapi_instance(self):
-        """✅ App should be FastAPI instance."""
-        from fastapi import FastAPI
-
-        assert isinstance(app, FastAPI)
-
-    def test_app_has_routes(self):
-        """✅ App should have routes registered."""
-        assert len(app.routes) > 0
-
-    def test_app_has_middleware(self):
-        """✅ App should have middleware."""
-        assert len(app.user_middleware) > 0
-
-    def test_app_from_settings(self):
-        """✅ App config should use settings."""
-        from app.core.config import settings
-
-        assert app.title == settings.APP_NAME
-        assert app.version == settings.APP_VERSION
-
-    def test_root_path_accessible(self):
-        """✅ Root path should be accessible."""
-        client = TestClient(app)
-
-        response = client.get("/")
-
-        # Should be accessible
-        assert response.status_code == 200
-
-    def test_docs_endpoint_accessible(self):
-        """✅ Docs endpoint should be accessible."""
-        client = TestClient(app)
-
-        response = client.get("/docs")
-
-        # Docs should be available (not 404)
-        assert response.status_code in [200, 307]
-
-
-class TestDatabaseInitialization:
-    """Test database initialization functions exist."""
-
-    def test_init_chromadb_imported(self):
-        """✅ init_chromadb should be imported in main."""
-        from app.main import init_chromadb
-
-        assert callable(init_chromadb)
-
-    def test_init_sqlite_imported(self):
-        """✅ init_sqlite should be imported in main."""
-        from app.main import init_sqlite
-
-        assert callable(init_sqlite)
-
-    def test_chromadb_returns_path(self):
-        """✅ init_chromadb should return a path."""
-        from pathlib import Path
-
-        from app.core.database import init_chromadb
-
-        result = init_chromadb()
-
-        # Should return a Path or string path
-        assert result is None or isinstance(result, (str, Path))
-
-    def test_sqlite_returns_url(self):
-        """✅ init_sqlite should return connection URL."""
-        from app.core.database import init_sqlite
-
-        result = init_sqlite()
-
-        # Should return a URL string
-        assert result is None or isinstance(result, str)
-
-
-class TestSettingsIntegration:
-    """Test settings are properly integrated."""
-
-    def test_settings_available_in_main(self):
-        """✅ Settings should be available in main module."""
-        from app.main import settings
-
-        assert settings is not None
-
-    def test_app_has_settings_title(self):
-        """✅ App title should match settings."""
-
-        assert app.title is not None
-        assert len(app.title) > 0
-
-
-class TestLoggerConfiguration:
-    """Test logging is configured."""
-
-    def test_logger_exists_in_main(self):
-        """✅ Logger should exist in main."""
-        from app.main import logger
-
-        assert logger is not None
-
-    def test_logger_has_name(self):
-        """✅ Logger should have module name."""
-        from app.main import logger
-
-        assert logger.name == "app.main"
-
-
-class TestCORSConfiguration:
-    """Test CORS is configured."""
-
-    def test_cors_middleware_exists(self):
-        """✅ CORS middleware should be registered."""
-        from app.main import app
-
-        middleware_names = [m.cls.__name__ for m in app.user_middleware]
-
-        assert "CORSMiddleware" in middleware_names
-
-    def test_api_v1_router_included(self):
-        """✅ API v1 router should be included."""
-        from app.main import app
-
-        routes = [route.path for route in app.routes]
-
-        # Should have some API v1 routes
-        has_api_v1 = any("/api/v1" in str(route) for route in routes)
-
-        assert has_api_v1 or len(routes) > 0  # At least has routes
-
-
-class TestApplicationState:
-    """Test application state and configuration."""
-
-    def test_app_has_description(self):
-        """✅ App should have description."""
-        assert app.description is not None
-
-    def test_app_debug_from_settings(self):
-        """✅ Debug mode should come from settings."""
-        from app.core.config import settings
-
-        assert app.debug == settings.DEBUG
-
-    def test_app_has_exception_handlers(self):
-        """✅ App should be able to handle exceptions."""
-        assert hasattr(app, "exception_handlers")
-
-
-class TestApiV1Integration:
-    """Test API v1 integration."""
-
-    def test_api_v1_router_accessible(self):
-        """✅ API v1 should be accessible."""
-        client = TestClient(app)
-
-        # Try root
-        response = client.get("/")
-
-        assert response.status_code == 200
-
-    def test_health_check_accessible(self):
-        """✅ Health check should be accessible."""
-        client = TestClient(app)
-
-        response = client.get("/api/v1/health")
-
-        # Should respond (not error 500)
-        assert response.status_code < 500
-
-    def test_openapi_schema_available(self):
-        """✅ OpenAPI schema should be available."""
-        client = TestClient(app)
-
-        response = client.get("/openapi.json")
-
-        # Should have schema
-        assert response.status_code == 200
