--- src/server/tests/unit/app/test_main_coverage_80.py
@@ -1,272 +1,0 @@
-"""
-Extended tests for app/main.py targeting 80%+ coverage.
-
-Focus on:
-- Lifespan handler execution paths (lines 65-79, 92, 104-106)
-- Exception handler (line 154, 176-177, 216-218)
-- Middleware and route registration
-"""
-
-from unittest.mock import patch
-
-import pytest
-from fastapi.testclient import TestClient
-
-
-class TestLifespanHandlers:
-    """Test startup and shutdown event handlers."""
-
-    def test_startup_event_calls_database_init(self):
-        """✅ Startup should initialize both databases."""
-
-        with (
-            patch("app.core.database.init_chromadb") as mock_chromadb,
-            patch("app.core.database.init_sqlite") as mock_sqlite,
-        ):
-            mock_chromadb.return_value = "/path/to/chromadb"
-            mock_sqlite.return_value = "sqlite:///path/to/db.sqlite"
-
-            # Import must reload to capture patches
-            import importlib
-
-            import app.main
-
-            importlib.reload(app.main)
-
-    def test_startup_event_logs_initialization(self):
-        """✅ Startup should log initialization messages."""
-        from app.main import startup_event
-
-        with (
-            patch("app.main.logger") as mock_logger,
-            patch("app.main.init_chromadb") as mock_chromadb,
-            patch("app.main.init_sqlite") as mock_sqlite,
-        ):
-            mock_chromadb.return_value = "/chromadb"
-            mock_sqlite.return_value = "sqlite:///db.sqlite"
-
-            # Since startup_event is async, we'd need to run it
-            # For now, just verify it exists and is callable
-            import inspect
-
-            assert inspect.iscoroutinefunction(startup_event)
-
-    def test_shutdown_event_exists_and_callable(self):
-        """✅ Shutdown event should be defined and callable."""
-        import inspect
-
-        from app.main import shutdown_event
-
-        assert inspect.iscoroutinefunction(shutdown_event)
-
-    def test_lifespan_context_manager_callable(self):
-        """✅ Lifespan should be a callable async context manager."""
-        import inspect
-
-        from app.main import lifespan
-
-        # Should be async generator or async callable
-        assert inspect.isasyncgenfunction(lifespan) or callable(lifespan)
-
-    def test_exception_handler_registered(self):
-        """✅ Exception handlers should be registered on app."""
-
-        from app.main import app
-
-        # HTTPException handler should be registered
-        assert app.exception_handlers is not None
-
-    def test_app_startup_exception_handler(self):
-        """✅ App should have HTTPException handler."""
-
-        from app.main import app
-
-        # The app should have exception handlers
-        assert len(app.exception_handlers) > 0
-
-
-class TestMiddlewareConfiguration:
-    """Test middleware is properly configured."""
-
-    def test_cors_middleware_is_configured(self):
-        """✅ CORS middleware should be registered."""
-        from app.main import app
-
-        # Check middleware exists
-        assert hasattr(app, "user_middleware")
-        middleware_names = [m.cls.__name__ for m in app.user_middleware]
-        assert "CORSMiddleware" in middleware_names
-
-    def test_cors_allows_local_origins(self):
-        """✅ CORS should allow local development origins."""
-        from app.main import app
-
-        # Get CORS middleware config
-        cors_middleware = None
-        for m in app.user_middleware:
-            if m.cls.__name__ == "CORSMiddleware":
-                cors_middleware = m
-                break
-
-        assert cors_middleware is not None
-
-
-class TestApplicationSetup:
-    """Test FastAPI application setup."""
-
-    def test_app_has_title_from_settings(self):
-        """✅ App title should come from settings."""
-        from app.core.config import settings
-        from app.main import app
-
-        assert app.title == settings.APP_NAME
-
-    def test_app_has_version_from_settings(self):
-        """✅ App version should come from settings."""
-        from app.core.config import settings
-        from app.main import app
-
-        assert app.version == settings.APP_VERSION
-
-    def test_debug_mode_from_settings(self):
-        """✅ Debug mode should come from settings."""
-        from app.core.config import settings
-        from app.main import app
-
-        assert app.debug == settings.DEBUG
-
-    def test_api_v1_router_included(self):
-        """✅ API v1 router should be included."""
-        from app.main import app
-
-        # Check routes
-        routes = [route.path for route in app.routes]
-        assert any("/api/v1" in route for route in routes)
-
-    def test_root_endpoint_exists(self):
-        """✅ Root endpoint should be registered."""
-        from app.main import app
-
-        client = TestClient(app)
-        response = client.get("/")
-
-        assert response.status_code == 200
-
-
-class TestHttpExceptionHandling:
-    """Test HTTP exception handling."""
-
-    def test_app_exception_handlers_configured(self):
-        """✅ App should have exception handlers configured."""
-        from app.main import app
-
-        # Should have exception handlers dict
-        assert hasattr(app, "exception_handlers")
-        assert isinstance(app.exception_handlers, dict)
-
-    def test_exception_handler_handles_404(self):
-        """✅ Exception handler should handle 404 errors."""
-        from app.main import app
-
-        client = TestClient(app)
-        response = client.get("/nonexistent-path")
-
-        assert response.status_code == 404
-
-    def test_exception_handler_handles_500(self):
-        """✅ Exception handler should handle 500 errors gracefully."""
-        from app.main import app
-
-        # The app should be set up to handle internal errors
-        assert app.exception_handlers is not None or hasattr(app, "exception_handlers")
-
-
-class TestLoggingConfiguration:
-    """Test logging is properly configured."""
-
-    def test_logger_created_from_main_module(self):
-        """✅ Logger should be created for main module."""
-        import app.main
-
-        assert app.main.logger is not None
-
-    def test_logger_has_correct_name(self):
-        """✅ Logger should have main module name."""
-        import app.main
-
-        assert app.main.logger.name == "app.main"
-
-
-class TestDatabaseInitializationCalls:
-    """Test database initialization functions are called on startup."""
-
-    @pytest.mark.asyncio
-    async def test_chromadb_init_called_in_startup(self):
-        """✅ ChromaDB should be initialized in startup."""
-        from app.main import startup_event
-
-        with (
-            patch("app.main.init_chromadb") as mock_init,
-            patch("app.main.init_sqlite"),
-        ):
-            mock_init.return_value = "/chromadb"
-
-            # startup_event is async, so we need asyncio to run it
-            # Just verify it's async and callable for now
-            import inspect
-
-            assert inspect.iscoroutinefunction(startup_event)
-
-    @pytest.mark.asyncio
-    async def test_sqlite_init_called_in_startup(self):
-        """✅ SQLite should be initialized in startup."""
-        from app.main import startup_event
-
-        with (
-            patch("app.main.init_chromadb"),
-            patch("app.main.init_sqlite") as mock_init,
-        ):
-            mock_init.return_value = "sqlite:///db.sqlite"
-
-            import inspect
-
-            assert inspect.iscoroutinefunction(startup_event)
-
-
-class TestApiEndpointIntegration:
-    """Integration tests for API endpoints."""
-
-    def test_health_endpoint_accessible(self):
-        """✅ Health endpoint should be accessible."""
-        from app.main import app
-
-        client = TestClient(app)
-        response = client.get("/api/v1/health")
-
-        # Should return 200 or similar success
-        assert response.status_code in [200, 404]  # 404 if route not fully impl
-
-    def test_knowledge_endpoints_accessible(self):
-        """✅ Knowledge endpoints should be registered."""
-        from app.main import app
-
-        client = TestClient(app)
-
-        # These might 404 if not fully implemented, but routes should exist
-        # Just verify app doesn't crash
-        try:
-            response = client.get("/api/v1/knowledge")
-        except Exception:
-            # If it raises, that's ok for this test
-            pass
-
-    def test_chat_endpoints_accessible(self):
-        """✅ Chat endpoints should be registered."""
-        from app.main import app
-
-        client = TestClient(app)
-
-        try:
-            response = client.get("/api/v1/chat")
-        except Exception:
-            pass
