--- src/server/tests/unit/scripts/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for scripts module."""
