--- src/server/tests/unit/services/__init__.py
@@ -1,1 +1,0 @@
-# Empty init file for services tests package
