+++ tests/python/unit/app/__init__.py
@@ -1,1 +1,3 @@
-# Tests for app module
+"""
+Tests: Unit tests for core business logic.
+"""
