+++ tests/python/unit/app/test_startup_handlers.py
